refactor [James]: refactor order route
